requirements.txt
python-telegram-bot==20.7
sqlalchemy
psycopg2-binary
